Improve precision for repeated functions (#1)
Before this we were accepting a repeat interval as an int.
When multiplfying this to calculate delay times the rounding
error would become multiplied.

+++ Apps/AudioAppTemplate/Source/TempoSynthesizerComponent.h
@@ -28,7 +28,7 @@ namespace AudioApp
         std::string content;
         std::atomic<bool> dirty;
 
-        void flash(int duration);
+        void flash(float duration);
         void updateLabel(uint8_t beat);
         void updateLabelColour(juce::Colour newColour);
